Support sign value to gauge charging current.
Category: POWER
Bug: None

+++ drivers/power/bq27541_fuelgauger.c
@@ -527,7 +527,7 @@ static struct platform_device this_device = {
 int battery_mvolts;
 int battery_capacity;
 int battery_temperature;
-int charge_current;
+short charge_current;
 
 static void msm_battery_update_psy_status(void)
 {
@@ -538,11 +538,11 @@ static void msm_battery_update_psy_status(void)
 	udelay(100);
 	battery_temperature = bq27541_get_battery_temperature();
 	udelay(100);
-	charge_current = bq27541_get_battery_current();
+	charge_current = (short) bq27541_get_battery_current();
 	udelay(100);
 
 	printk("%s mvolts=%d capacity=%d temperature=%d charge_current=%d\n", __FUNCTION__,
-		(int) battery_mvolts, (int) battery_capacity, (int) battery_temperature, (int) charge_current);
+		(int) battery_mvolts, (int) battery_capacity, (int) battery_temperature, (short) charge_current);
 
 	queue_delayed_work(bq27541_di->battery_queue, &(bq27541_di->battery_work), BATT_POLLING_TIME);
 }
